FELIX-4888 : ServletHandler's are not sorted by longest matching path. DTO handling (WiP)
git-svn-id: https://svn.apache.org/repos/asf/felix/trunk@1680059 13f79535-47bb-0310-9956-ffa450edef68

+++ http/base/src/main/java/org/apache/felix/http/base/internal/runtime/dto/state/FailureFilterState.java
@@ -21,5 +21,5 @@ package org.apache.felix.http.base.internal.runtime.dto.state;
 
 public interface FailureFilterState extends FilterState {
 
-    long getReason();
+    int getReason();
 }
